Almost there

+++ Userland/SampleCodeModule/sampleCodeModule.c
@@ -1,9 +1,12 @@
 #include <userlib.h>
 #include <shell.h>
+#include <processlib.h>
+#include <Testing/test_util.h>
+#include <Testing/tests.h>
 
 int main() {
 
-    spawn_process(current_pid(), &test_processes, 0, NULL, 1, true);
+    spawn_process(current_pid(), (uint64_t) &test_processes, 0, NULL, 1, true);
     // spawn_process(current_pid(), &test_prio, 0, NULL, 1, true);
